Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-insert-expected.txt


FAIL Initial color-scheme assert_equals: expected "normal" but got "auto"
FAIL Inserted meta color-scheme applies assert_equals: expected "light" but got "auto"
FAIL Inserted meta color-scheme before existing in head applies assert_equals: expected "dark" but got "auto"
